+++ src/decoders.h
@@ -6,10 +6,10 @@
 
 #include "stream.h"
 
-void streamReadTag2_3S32(mmapStream_t *stream, int32_t *values);
-void streamReadTag8_4S16_v1(mmapStream_t *stream, int32_t *values);
-void streamReadTag8_4S16_v2(mmapStream_t *stream, int32_t *values);
-void streamReadTag8_8SVB(mmapStream_t *stream, int32_t *values, int valueCount);
+void streamReadTag2_3S32(mmapStream_t *stream, int64_t *values);
+void streamReadTag8_4S16_v1(mmapStream_t *stream, int64_t *values);
+void streamReadTag8_4S16_v2(mmapStream_t *stream, int64_t *values);
+void streamReadTag8_8SVB(mmapStream_t *stream, int64_t *values, int valueCount);
 
 int16_t streamReadS16(mmapStream_t *stream);
