+++ src/gpxwriter.c
@@ -27,7 +27,7 @@ void gpxWriterAddPreamble(gpxWriter_t *gpx)
  * Time is in microseconds since device power-on. Lat and lon are degrees multiplied by GPS_DEGREES_DIVIDER. Altitude
  * is in meters.
  */
-void gpxWriterAddPoint(gpxWriter_t *gpx, uint32_t time, int32_t lat, int32_t lon, int16_t altitude)
+void gpxWriterAddPoint(gpxWriter_t *gpx, int64_t time, int32_t lat, int32_t lon, int16_t altitude)
 {
     char negSign[] = "-";
     char noSign[] = "";
@@ -53,10 +53,10 @@ void gpxWriterAddPoint(gpxWriter_t *gpx, uint32_t time, int32_t lat, int32_t lon
     char *lonSign = ((lon < 0) && (lonDegrees == 0)) ? negSign : noSign;
 
     fprintf(gpx->file, "  <trkpt lat=\"%s%d.%07u\" lon=\"%s%d.%07u\"><ele>%d</ele>", latSign, latDegrees, latFracDegrees, lonSign, lonDegrees, lonFracDegrees, altitude);
-    
-    if (time != (uint32_t) -1) {
+
+    if (time != -1) {
         //We'll just assume that the timespan is less than 24 hours, and make up a date
-        int hours, mins, secs, frac;
+        uint32_t hours, mins, secs, frac;
 
         frac = time % 1000000;
         secs = time / 1000000;
@@ -67,7 +67,7 @@ void gpxWriterAddPoint(gpxWriter_t *gpx, uint32_t time, int32_t lat, int32_t lon
         hours = mins / 60;
         mins %= 60;
 
-        fprintf(gpx->file, "<time>2000-01-01T%02d:%02d:%02d.%06uZ</time>", hours, mins, secs, frac);
+        fprintf(gpx->file, "<time>2000-01-01T%02u:%02u:%02u.%06uZ</time>", hours, mins, secs, frac);
     }
     fprintf(gpx->file, "</trkpt>\n");
 }
